+++ database/database.sql
@@ -1,0 +1,5 @@
+CREATE DATABASE IF NOT EXISTS proud_peach
+  CHARACTER SET utf8mb4
+  COLLATE utf8mb4_unicode_ci;
+
+USE proud_peach;