+++ changelog.md
@@ -31,3 +31,6 @@ Principles of a change log, excerpted from [Keep a Changelog](http://keepachange
  * License, git metafiles and ReadMe file.
  * Changelog file.
  * Metadata YAML file.
+ * Base site design and HTML.
+ * Database schematics.
+ * ICD-10 list.